diplome-generator/requirements.txt
ezodf
pandas
python-pptx
pyinstaller
